fix:进行了数据存储规范，将数据全部存储至结构体内以方便会调函数使用
