+++ src/auth.c
@@ -5,6 +5,7 @@
 #include <stdlib.h>
 int login_verify(db* _database, str username, str password){
     int line_no = database_query_by_column(_database, "account", "username", username);
+    perror("\n");
     if(line_no == -1) return -1;
     if(strcmp(password, database_select(_database, "account", "password", line_no)) == 0) {
         return atoi(database_select(_database, "account", "auth", line_no));
